EMPIAR reader as intake Catalog/Datasource
